Initial Commit, containing the older build.
This build was made in Feb of 2020, when the team worked together last time on this project. Later on the finalization was postponed for a while. This is the Alpha test build for showing the progress the team made at that time. (HOLESOME WORLD!)
--ProjectTeam-- 
Gann4Life
Pix_Aman
Greg97
Jimm